+++ src/views/setting/nav.php
@@ -1,20 +1,19 @@
 <?php declare(strict_types=1);
 
 use Dwnload\WpSettingsApi\Api\SettingSection;
-use Dwnload\WpSettingsApi\App;
 use Dwnload\WpSettingsApi\Settings\SectionManager;
+use Dwnload\WpSettingsApi\WpSettingsApi;
 
 /** @var $this Dwnload\WpSettingsApi\WpSettingsApi */
-$sections = SectionManager::getSections($this->getApp()->getMenuSlug());
+$sections = SectionManager::getSection($this->getPluginInfo()->getMenuSlug());
 
-add_action(App::ACTION_PREFIX . 'settings_sidebars', function () {
+add_action(WpSettingsApi::ACTION_PREFIX . 'settings_sidebars', function () {
     echo '<ul class="Dwnload_WP_Settings_Api__menu">';
 }, 0);
 
-/** @var SettingSection $section */
 foreach ($sections as $key => $section) {
     ++$key;
-    add_action(App::ACTION_PREFIX . 'settings_sidebars', function () use ($section) {
+    add_action(WpSettingsApi::ACTION_PREFIX . 'settings_sidebars', function () use ($section) {
         printf(
             '<li><a href="javascript:;" data-tab-id="Dwnload_WP_Settings_Api__%s">%s</a></li>',
             esc_attr($section->getId()),
@@ -23,6 +22,6 @@ foreach ($sections as $key => $section) {
     }, ++$key * 2 + 1);
 }
 
-add_action(App::ACTION_PREFIX . 'settings_sidebars', function () {
+add_action(WpSettingsApi::ACTION_PREFIX . 'settings_sidebars', function () {
     echo '</ul>';
 }, 199);